slackrest/requirements.txt
behave
tornado
slackclient